% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertplot}
\alias{convertplot}
\title{convert basic plot to ggplot object}
\usage{
convertplot(exp)
}
\arguments{
\item{exp}{expression formula of basic plot}
}
\value{
ggplot object
}
\description{
convert basic plot to ggplot object
}
\examples{
\donttest{
p <- convertplot(plot(1:10))
class(p)
print(p)
}
}
\author{
Kai Guo
}
