% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.setcolnames}
\alias{.setcolnames}
\title{set first row as column name}
\usage{
.setcolnames(file)
}
\arguments{
\item{file}{input data frame}
}
\description{
set first row as column name
}
\author{
Kai Guo
}
