% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inoffice.R
\name{inpptx}
\alias{inpptx}
\title{read table from pptx}
\usage{
inpptx(filename, header = FALSE)
}
\arguments{
\item{filename}{input filename}

\item{header}{use first row as column name}
}
\description{
read table from pptx
}
\examples{
\dontrun{
totable(t.test(wt ~ am, mtcars), filename = "mtcars.pptx")
tabs <- inpptx(filename = "mtcars.pptx", header = TRUE)
tabs
}
}
\author{
Kai Guo
}
