% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{psmelt}
\alias{psmelt}
\title{Melt phyloseq data object into large data.frame}
\usage{
psmelt(physeq, ...)
}
\arguments{
\item{physeq}{A sample_data-class, or a phyloseq-class object with a sample_data.
If the sample_data slot is missing in physeq, then physeq will be returned as-is,
and a warning will be printed to screen.}

\item{...}{parameters for the subset_samples function in phyloseq package}
}
\description{
Melt phyloseq data object into large data.frame
}
