% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{subset_samples}
\alias{subset_samples}
\title{Subset the phyloseq based on sample}
\usage{
subset_samples(physeq, ...)
}
\arguments{
\item{physeq}{A sample_data-class, or a phyloseq-class object with a sample_data.
If the sample_data slot is missing in physeq, then physeq will be returned as-is,
and a warning will be printed to screen.}

\item{...}{parameters for the subset_samples function in phyloseq package}
}
\description{
Subset the phyloseq based on sample
}
