% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tax_table}
\alias{tax_table}
\title{extract taxonomy table}
\usage{
tax_table(physeq, ...)
}
\arguments{
\item{physeq}{An object among the set of classes defined by the phyloseq package that contain taxonomyTable.}

\item{...}{parameters for the tax_table function in phyloseq package}
}
\description{
extract taxonomy table
}
