% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toFigure.R
\name{tofigure}
\alias{tofigure}
\title{output figures to different formats}
\usage{
tofigure(figure, format = NULL, filename = "temp.pdf")
}
\arguments{
\item{figure}{output figure function, set NULL output the current figure}

\item{type}{output format}
}
\description{
output figures to different formats
}
\author{
Kai Guo
}
