% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tohtml.R
\name{tohtml}
\alias{tohtml}
\title{export as plotly html (only support ggplot2 object)}
\usage{
tohtml(figure, filename = "temp.html", save = TRUE)
}
\arguments{
\item{figure}{output figure function, set NULL output the current figure}

\item{filename}{output format (could be ingored)}

\item{save}{save figure or not (default: TRUE)}
}
\description{
export as plotly html (only support ggplot2 object)
}
\examples{
\dontrun{
##set tempdir used for the examples
currentwd <- getwd()
setwd(tempdir())
p <- ggplot(mtcars, aes(mpg, disp, color = factor(cyl))) + geom_point()
tohtml(p,filename = "mtcars.html")
## or use ggplot directly
tohtml(ggplot(mtcars, aes(mpg, disp, color = factor(cyl))) +
geom_point(), filename = "mtcars.html")
setwd(currentwd)
}
}
\author{
Kai Guo
}
