% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTable.R
\name{totable}
\alias{totable}
\title{write table out to pptx or docx}
\usage{
totable(data, filename, format = NULL, append = FALSE)
}
\arguments{
\item{data}{datasets}

\item{filename}{output filename}

\item{format}{pptx or docx}

\item{append}{append into file}
}
\description{
write table out to pptx or docx
}
\examples{
\dontrun{
## set tempdir used for the examples
currentwd <- getwd()
setwd(tempdir())
tt <- t.test(wt ~ am, mtcars)
totable(tt, filename = "mtcars_test.pptx")
totable(t.test(wt ~ am, mtcars), filename = "mtcars_test.pptx")
totable(head(mtcars), filename = "mtcars_head.docx")
## set back to the previous work directory
setwd(currentwd)
}
}
\author{
Kai Guo
}
