% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{convertplot}
\alias{convertplot}
\title{convert basic plot to ggplot object}
\usage{
convertplot(exp)
}
\arguments{
\item{expr}{expression formula of basic plot}
}
\value{
ggplot object
}
\description{
convert basic plot to ggplot object
}
\author{
Kai Guo
}
