% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inoffice.R
\name{indocx}
\alias{indocx}
\title{read table from docx}
\usage{
indocx(filename, header = FALSE)
}
\arguments{
\item{filename}{input filename}

\item{header}{use first row as column name}
}
\description{
read table from docx
}
\examples{
totable(t.test(wt ~ am, mtcars), filename = file.path(tempdir(), "mtcars.docx"))
tabs <- indocx(filename = file.path(tempdir(), "mtcars.docx"), header = TRUE)
tabs
}
\author{
Kai Guo
}
