% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inoffice.R
\name{inoffice}
\alias{inoffice}
\title{read tables from ppt and word}
\usage{
inoffice(filename, format = NULL, header = TRUE)
}
\arguments{
\item{filename}{input filename}

\item{format}{pptx or docx}

\item{header}{use first row as column name}
}
\description{
read tables from ppt and word
}
\examples{
## use tempdir for the example
totable(t.test(wt ~ am, mtcars), filename = file.path(tempdir(), "mtcars.pptx"))
tabs <- inoffice(filename = file.path(tempdir(), "mtcars.pptx"), header = TRUE)
tabs
}
\author{
Kai Guo
}
