% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{todocx}
\alias{todocx}
\title{export figure to docx}
\usage{
todocx(figure = NULL, filename = NULL, width = 6, height = 6,
  append = FALSE)
}
\arguments{
\item{figure}{plot figure function}

\item{filename}{output filename}

\item{width}{width of the output figure}

\item{height}{height of the output figure}

\item{append}{append or not}
}
\description{
export figure to docx
}
\examples{
if(interactive()){
plot(mtcars$mpg, mtcars$disp)
todocx(filename = file.path(tempdir(), "mtcars.docx"))
## use ggplot2
ggplot(mtcars, aes(mpg, disp, color = factor(cyl))) + geom_point()
todocx(filename = file.path(tempdir(), "mtcars.docx"), height = 6, width = 4)
}
}
\author{
Kai Guo
}
