% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{todocx}
\alias{todocx}
\title{export figure to docx}
\usage{
todocx(figure = NULL, filename = NULL, append = FALSE)
}
\arguments{
\item{figure}{figure function}

\item{filename}{output filename}
}
\description{
export figure to docx
}
\author{
Kai Guo
}
