% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{todocx}
\alias{todocx}
\title{export figure to docx}
\usage{
todocx(figure = NULL, filename = NULL, width = 6, height = 6,
  append = FALSE)
}
\arguments{
\item{figure}{plot figure function}

\item{filename}{output filename}

\item{width}{width of the output figure}

\item{height}{height of the output figure}

\item{append}{append or not}
}
\description{
export figure to docx
}
\examples{
\dontrun{
## set tempdir used for the examples
currentwd <- getwd()
setwd(tempdir())
plot(mtcars$mpg, mtcars$disp)
todocx(filename = "mtcars.docx")
## You can also use ggplot
ggplot(mtcars, aes(mpg, disp, color = factor(cyl))) + geom_point()
todocx(filename = "mtcars.docx", height = 6, width = 4)
## set back to the previous work directory
setwd(currentwd)
}
}
\author{
Kai Guo
}
