% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{toffice}
\alias{toffice}
\title{export graph to MS office}
\usage{
toffice(figure = NULL, format = "pptx", filename = "temp.pptx",
  append = FALSE, width = 4, height = 4)
}
\arguments{
\item{figure}{plot figure function}

\item{format}{file format}

\item{filename}{output filename}

\item{append}{append or not}

\item{width}{width of the output figure}

\item{height}{height of the output figure}
}
\description{
export graph to MS office
}
\examples{
\dontrun{
## set tempdir used for the examples
currentwd <- getwd()
setwd(tempdir())
plot(mtcars$mpg, mtcars$disp)
toffice(filename = "mtcars.pptx",format = "pptx)
## You can also use ggplot
ggplot(mtcars, aes(mpg, disp, color = factor(cyl))) + geom_point()
toffice(filename = "mtcars.pptx",format = "pptx)
## set back to the previous work directory
setwd(currentwd)
}
}
\author{
Kai Guo
}
