% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{toffice}
\alias{toffice}
\title{export graph to MS office}
\usage{
toffice(figure = NULL, format = "pptx", filename = "temp.pptx",
  append = FALSE, width = 4, height = 4)
}
\arguments{
\item{figure}{figure function}

\item{format}{file type}

\item{filename}{output filename}

\item{append}{append into file}

\item{width}{width of the output figure}

\item{height}{height of the output figure}
}
\description{
export graph to MS office
}
\author{
Kai Guo
}
