% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{topptx}
\alias{topptx}
\title{export figure to pptx}
\usage{
topptx(figure = NULL, filename = NULL, append = FALSE, width = 4,
  height = 4)
}
\arguments{
\item{figure}{figure function}

\item{filename}{output filename}

\item{append}{append into file}

\item{width}{width of the output figure}

\item{height}{height of the output figure}
}
\description{
export figure to pptx
}
\author{
Kai Guo
}
