% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toffice.R
\name{topptx}
\alias{topptx}
\title{export figure to pptx}
\usage{
topptx(figure = NULL, filename = NULL, append = FALSE)
}
\arguments{
\item{figure}{figure function}

\item{filename}{output filename}
}
\description{
export figure to pptx
}
\author{
Kai Guo
}
