% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTable.R
\name{toTable}
\alias{toTable}
\title{write table out to pptx or docx}
\usage{
toTable(data, filename, format = NULL)
}
\arguments{
\item{data}{datasets}

\item{filename}{output filename}
}
\description{
write table out to pptx or docx
}
\author{
Kai Guo
}
