% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTable.R
\name{totable}
\alias{totable}
\title{write table out to pptx or docx}
\usage{
totable(data, filename, format = NULL, append = FALSE)
}
\arguments{
\item{data}{datasets}

\item{filename}{output filename}

\item{format}{pptx or docx}

\item{append}{append into file}
}
\description{
write table out to pptx or docx
}
\author{
Kai Guo
}
